+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ### New features
  - add `.docx` files for installation guides
 
+### Bugfix
+ - properly load `hooks` in cronjobs
+ 
+
 # v6.9.0
 ### New features
  - set field type for field `count` in style `progressBar` to `text`. This way the field can be dynamically loaded with interpolation. It still should be a number otherwise it is assigned to `0`